+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed `pac_lifeduration`, not work for old packages which do not have UTC in published Date.
 * `checkred` argument from `lib_validate` will expecting any values from `c("ERROR", "FAIL", "WARN", "NOTE")` vector.
+* Default `scope` for `pac_checkred` will be `c("ERROR", "FAIL")`.
 
 # 0.2.8